+++ vitest.config.ts
@@ -4,6 +4,6 @@ export default defineConfig({
   test: {
     environment: "jsdom",
     setupFiles: ["./vitest.setup.ts"],
-    css: true
-  }
+    css: true,
+  },
 });